Use default text placement fns
